renaming package to demand sdk adapters
