+++ PrebidMobile/DemandSDKAdaptersForDFP/build.gradle
