+++ PrebidMobile/DemandSDKAdaptersForDFP/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest package="org.prebid.demandsdkadapters.dfp" />
+
