+++ PrebidMobile/DemandSDKAdaptersForDFP/src/main/java/org/prebid/demandsdkadapters/dfp/PrebidCustomEventInterstitial.java
@@ -1,10 +1,13 @@
-package org.prebid.mediationadapters.mopub;
+package org.prebid.demandsdkadapters.dfp;
 
 
 import android.content.Context;
+import android.os.Bundle;
 
-import com.mopub.mobileads.CustomEventInterstitial;
-import com.mopub.mobileads.MoPubErrorCode;
+import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.mediation.MediationAdRequest;
+import com.google.android.gms.ads.mediation.customevent.CustomEventInterstitial;
+import com.google.android.gms.ads.mediation.customevent.CustomEventInterstitialListener;
 
 import org.json.JSONObject;
 import org.prebid.mobile.prebidserver.CacheService;
@@ -13,48 +16,45 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
 import java.lang.reflect.Proxy;
-import java.util.Map;
 
-import static org.prebid.mediationadapters.mopub.PrebidCustomEventSettings.*;
+import static org.prebid.demandsdkadapters.dfp.PrebidCustomEventSettings.*;
 
-public class PrebidCustomEventInterstitial extends CustomEventInterstitial {
-    private String bidder;
+public class PrebidCustomEventInterstitial implements CustomEventInterstitial {
+    private String bidderName;
     private Object adObject;
 
     @Override
-    protected void loadInterstitial(Context context, CustomEventInterstitialListener customEventInterstitialListener, Map<String, Object> localExtras, Map<String, String> serverExtras) {
-        if (localExtras != null) {
-            String cache_id = (String) localExtras.get(PREBID_CACHE_ID);
-            bidder = (String) localExtras.get(PREBID_BIDDER);
-            if (FACEBOOK_BIDDER_NAME.equals(bidder) && demandSet.contains(PrebidCustomEventSettings.Demand.FACEBOOK)) {
-//            if ("appnexus".equals(bidder)) {
-                loadFacebookInterstitial(context, cache_id, customEventInterstitialListener);
+    public void requestInterstitialAd(Context context, CustomEventInterstitialListener customEventInterstitialListener, String s, MediationAdRequest mediationAdRequest, Bundle bundle) {
+        if (bundle != null) {
+            String cacheId = (String) bundle.get(PREBID_CACHE_ID);
+            bidderName = (String) bundle.get(PREBID_BIDDER);
+            if (FACEBOOK_BIDDER_NAME.equals(bidderName) && demandSet.contains(PrebidCustomEventSettings.Demand.FACEBOOK)) {
+                loadFacebookInterstitial(context, cacheId, customEventInterstitialListener);
             } else {
                 if (customEventInterstitialListener != null) {
-                    customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.NO_FILL);
+                    customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_NO_FILL);
                 }
             }
         } else {
             if (customEventInterstitialListener != null) {
-                customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.INTERNAL_ERROR);
+                customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INVALID_REQUEST);
             }
         }
     }
 
     @Override
-    protected void showInterstitial() {
-        if (FACEBOOK_BIDDER_NAME.equals(bidder)) {
+    public void showInterstitial() {
+        if (FACEBOOK_BIDDER_NAME.equals(bidderName)) {
             try {
                 adObject.getClass().getMethod(FACEBOOK_SHOW_METHOD).invoke(adObject);
             } catch (Exception e) {
             }
         }
-
     }
 
     @Override
-    protected void onInvalidate() {
-        if (FACEBOOK_BIDDER_NAME.equals(bidder)) {
+    public void onDestroy() {
+        if (FACEBOOK_BIDDER_NAME.equals(bidderName)) {
             try {
                 adObject.getClass().getMethod(FACEBOOK_DESTROY_METHOD).invoke(adObject);
             } catch (Exception e) {
@@ -62,11 +62,22 @@ public class PrebidCustomEventInterstitial extends CustomEventInterstitial {
         }
     }
 
+    @Override
+    public void onPause() {
+
+    }
+
+    @Override
+    public void onResume() {
+
+    }
+
     private void loadFacebookInterstitial(final Context context, String cacheId, final CustomEventInterstitialListener customEventInterstitialListener) {
         CacheService cs = new CacheService(new CacheService.CacheListener() {
             @Override
             public void onResponded(JSONObject jsonObject) {
                 try {
+
                     String adm = jsonObject.getString(PREBID_ADM);
                     JSONObject bid = new JSONObject(adm);
                     String placementID = bid.getString(FACEBOOK_PLACEMENT_ID);
@@ -78,50 +89,51 @@ public class PrebidCustomEventInterstitial extends CustomEventInterstitial {
                         @Override
                         public Object invoke(Object o, Method method, Object[] objects) throws Throwable {
                             if (method.getName().equals(FACEBOOK_ON_INTERSTITIAL_DISPLAYED_METHOD)) {
-                                customEventInterstitialListener.onInterstitialShown();
+                                customEventInterstitialListener.onAdOpened();
                             } else if (method.getName().equals(FACEBOOK_ON_INTERSTITIAL_DISMISSED_METHOD)) {
-                                customEventInterstitialListener.onInterstitialDismissed();
+                                customEventInterstitialListener.onAdClosed();
                             } else if (method.getName().equals(FACEBOOK_ON_ERROR_METHOD)) {
                                 Method getErrorCode = objects[1].getClass().getMethod(FACEBOOK_GET_ERROR_CODE_METHOD);
                                 switch ((int) getErrorCode.invoke(objects[1])) {
                                     case 1000:
-                                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.NO_CONNECTION);
+                                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_NETWORK_ERROR);
                                         break;
                                     case 1001:
-                                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.NO_FILL);
+                                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_NO_FILL);
                                         break;
                                     case 1002:
-                                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.INTERNAL_ERROR);
+                                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INVALID_REQUEST);
                                         break;
                                     case 2000:
-                                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.SERVER_ERROR);
+                                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INTERNAL_ERROR);
                                         break;
                                     case 2001:
-                                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.INTERNAL_ERROR);
+                                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INTERNAL_ERROR);
                                         break;
                                     case 3001:
-                                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.INTERNAL_ERROR);
+                                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INTERNAL_ERROR);
                                         break;
                                 }
                             } else if (method.getName().equals(FACEBOOK_ON_AD_LOADED_METHOD)) {
-                                customEventInterstitialListener.onInterstitialLoaded();
+                                customEventInterstitialListener.onAdLoaded();
                             } else if (method.getName().equals(FACEBOOK_ON_AD_CLICKED_METHOD)) {
-                                customEventInterstitialListener.onInterstitialClicked();
+                                customEventInterstitialListener.onAdClicked();
                             }
                             return null;
                         }
                     });
-                    Method setAdListener = interstitialClass.getMethod("setAdListener", adLisenterInterface);
+                    Method setAdListener = interstitialClass.getMethod(FACEBOOK_SET_AD_LISTENER_METHOD, adLisenterInterface);
                     setAdListener.invoke(adObject, newAdListener);
-                    Method loadAdFromBid = interstitialClass.getMethod("loadAdFromBid", String.class);
+                    Method loadAdFromBid = interstitialClass.getMethod(FACEBOOK_LOAD_AD_FROM_BID_METHOD, String.class);
                     loadAdFromBid.invoke(adObject, adm);
                 } catch (Exception e) {
                     if (customEventInterstitialListener != null) {
-                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.INTERNAL_ERROR);
+                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INTERNAL_ERROR);
                     }
                 }
             }
         }, cacheId);
         cs.execute();
+
     }
 }
