+++ PrebidMobile/DemandSDKAdaptersForMoPub/src/main/AndroidManifest.xml
@@ -1,0 +1,3 @@
+<manifest package="org.prebid.demandsdkadapters.mopub" />
+
+
