+++ PrebidMobile/DemandSDKAdaptersForMoPub/src/main/java/org/prebid/demandsdkadapters/mopub/PrebidCustomEventInterstitial.java
@@ -1,13 +1,10 @@
-package org.prebid.mediationadapters.dfp;
+package org.prebid.demandsdkadapters.mopub;
 
 
 import android.content.Context;
-import android.os.Bundle;
 
-import com.google.android.gms.ads.AdRequest;
-import com.google.android.gms.ads.mediation.MediationAdRequest;
-import com.google.android.gms.ads.mediation.customevent.CustomEventInterstitial;
-import com.google.android.gms.ads.mediation.customevent.CustomEventInterstitialListener;
+import com.mopub.mobileads.CustomEventInterstitial;
+import com.mopub.mobileads.MoPubErrorCode;
 
 import org.json.JSONObject;
 import org.prebid.mobile.prebidserver.CacheService;
@@ -16,45 +13,48 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
 import java.lang.reflect.Proxy;
+import java.util.Map;
 
-import static org.prebid.mediationadapters.dfp.PrebidCustomEventSettings.*;
+import static org.prebid.demandsdkadapters.mopub.PrebidCustomEventSettings.*;
 
-public class PrebidCustomEventInterstitial implements CustomEventInterstitial {
-    private String bidderName;
+public class PrebidCustomEventInterstitial extends CustomEventInterstitial {
+    private String bidder;
     private Object adObject;
 
     @Override
-    public void requestInterstitialAd(Context context, CustomEventInterstitialListener customEventInterstitialListener, String s, MediationAdRequest mediationAdRequest, Bundle bundle) {
-        if (bundle != null) {
-            String cacheId = (String) bundle.get(PREBID_CACHE_ID);
-            bidderName = (String) bundle.get(PREBID_BIDDER);
-            if (FACEBOOK_BIDDER_NAME.equals(bidderName) && demandSet.contains(PrebidCustomEventSettings.Demand.FACEBOOK)) {
-                loadFacebookInterstitial(context, cacheId, customEventInterstitialListener);
+    protected void loadInterstitial(Context context, CustomEventInterstitialListener customEventInterstitialListener, Map<String, Object> localExtras, Map<String, String> serverExtras) {
+        if (localExtras != null) {
+            String cache_id = (String) localExtras.get(PREBID_CACHE_ID);
+            bidder = (String) localExtras.get(PREBID_BIDDER);
+            if (FACEBOOK_BIDDER_NAME.equals(bidder) && demandSet.contains(PrebidCustomEventSettings.Demand.FACEBOOK)) {
+//            if ("appnexus".equals(bidder)) {
+                loadFacebookInterstitial(context, cache_id, customEventInterstitialListener);
             } else {
                 if (customEventInterstitialListener != null) {
-                    customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_NO_FILL);
+                    customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.NO_FILL);
                 }
             }
         } else {
             if (customEventInterstitialListener != null) {
-                customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INVALID_REQUEST);
+                customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.INTERNAL_ERROR);
             }
         }
     }
 
     @Override
-    public void showInterstitial() {
-        if (FACEBOOK_BIDDER_NAME.equals(bidderName)) {
+    protected void showInterstitial() {
+        if (FACEBOOK_BIDDER_NAME.equals(bidder)) {
             try {
                 adObject.getClass().getMethod(FACEBOOK_SHOW_METHOD).invoke(adObject);
             } catch (Exception e) {
             }
         }
+
     }
 
     @Override
-    public void onDestroy() {
-        if (FACEBOOK_BIDDER_NAME.equals(bidderName)) {
+    protected void onInvalidate() {
+        if (FACEBOOK_BIDDER_NAME.equals(bidder)) {
             try {
                 adObject.getClass().getMethod(FACEBOOK_DESTROY_METHOD).invoke(adObject);
             } catch (Exception e) {
@@ -62,22 +62,11 @@ public class PrebidCustomEventInterstitial implements CustomEventInterstitial {
         }
     }
 
-    @Override
-    public void onPause() {
-
-    }
-
-    @Override
-    public void onResume() {
-
-    }
-
     private void loadFacebookInterstitial(final Context context, String cacheId, final CustomEventInterstitialListener customEventInterstitialListener) {
         CacheService cs = new CacheService(new CacheService.CacheListener() {
             @Override
             public void onResponded(JSONObject jsonObject) {
                 try {
-
                     String adm = jsonObject.getString(PREBID_ADM);
                     JSONObject bid = new JSONObject(adm);
                     String placementID = bid.getString(FACEBOOK_PLACEMENT_ID);
@@ -89,51 +78,50 @@ public class PrebidCustomEventInterstitial implements CustomEventInterstitial {
                         @Override
                         public Object invoke(Object o, Method method, Object[] objects) throws Throwable {
                             if (method.getName().equals(FACEBOOK_ON_INTERSTITIAL_DISPLAYED_METHOD)) {
-                                customEventInterstitialListener.onAdOpened();
+                                customEventInterstitialListener.onInterstitialShown();
                             } else if (method.getName().equals(FACEBOOK_ON_INTERSTITIAL_DISMISSED_METHOD)) {
-                                customEventInterstitialListener.onAdClosed();
+                                customEventInterstitialListener.onInterstitialDismissed();
                             } else if (method.getName().equals(FACEBOOK_ON_ERROR_METHOD)) {
                                 Method getErrorCode = objects[1].getClass().getMethod(FACEBOOK_GET_ERROR_CODE_METHOD);
                                 switch ((int) getErrorCode.invoke(objects[1])) {
                                     case 1000:
-                                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_NETWORK_ERROR);
+                                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.NO_CONNECTION);
                                         break;
                                     case 1001:
-                                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_NO_FILL);
+                                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.NO_FILL);
                                         break;
                                     case 1002:
-                                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INVALID_REQUEST);
+                                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.INTERNAL_ERROR);
                                         break;
                                     case 2000:
-                                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INTERNAL_ERROR);
+                                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.SERVER_ERROR);
                                         break;
                                     case 2001:
-                                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INTERNAL_ERROR);
+                                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.INTERNAL_ERROR);
                                         break;
                                     case 3001:
-                                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INTERNAL_ERROR);
+                                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.INTERNAL_ERROR);
                                         break;
                                 }
                             } else if (method.getName().equals(FACEBOOK_ON_AD_LOADED_METHOD)) {
-                                customEventInterstitialListener.onAdLoaded();
+                                customEventInterstitialListener.onInterstitialLoaded();
                             } else if (method.getName().equals(FACEBOOK_ON_AD_CLICKED_METHOD)) {
-                                customEventInterstitialListener.onAdClicked();
+                                customEventInterstitialListener.onInterstitialClicked();
                             }
                             return null;
                         }
                     });
-                    Method setAdListener = interstitialClass.getMethod(FACEBOOK_SET_AD_LISTENER_METHOD, adLisenterInterface);
+                    Method setAdListener = interstitialClass.getMethod("setAdListener", adLisenterInterface);
                     setAdListener.invoke(adObject, newAdListener);
-                    Method loadAdFromBid = interstitialClass.getMethod(FACEBOOK_LOAD_AD_FROM_BID_METHOD, String.class);
+                    Method loadAdFromBid = interstitialClass.getMethod("loadAdFromBid", String.class);
                     loadAdFromBid.invoke(adObject, adm);
                 } catch (Exception e) {
                     if (customEventInterstitialListener != null) {
-                        customEventInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INTERNAL_ERROR);
+                        customEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.INTERNAL_ERROR);
                     }
                 }
             }
         }, cacheId);
         cs.execute();
-
     }
 }
