+++ PrebidMobile/DemandSDKAdaptersForMoPub/src/main/java/org/prebid/demandsdkadapters/mopub/PrebidCustomEventSettings.java
@@ -1,4 +1,4 @@
-package org.prebid.mediationadapters.mopub;
+package org.prebid.demandsdkadapters.mopub;
 
 import android.content.Context;
