+++ PrebidMobile/DemoApp/build.gradle
@@ -35,7 +35,7 @@ dependencies {
     }
     compile project(":PrebidMobileCore")
     compile project(":PrebidServerAdapter")
-    compile project(":MediationAdaptersForDFP")
-    compile project(":MediationAdaptersForMoPub")
+    compile project(":DemandSDKAdaptersForDFP")
+    compile project(":DemandSDKAdaptersForMoPub")
     compile(name: 'ads-release', ext: 'aar')
 }
