--- PrebidMobile/MediationAdaptersForDFP/src/main/AndroidManifest.xml
@@ -1,2 +1,0 @@
-<manifest package="org.prebid.mediationadapters.dfp" />
-
