--- PrebidMobile/MediationAdaptersForMoPub/src/main/AndroidManifest.xml
@@ -1,3 +1,0 @@
-<manifest package="org.prebid.mediationadapters.mopub" />
-
-
