+++ PrebidMobile/settings.gradle
@@ -2,5 +2,5 @@ include ':DemoApp'
 include ':PrebidMobileCore'
 include ':UnitTestUtils'
 include ':PrebidServerAdapter'
-include ':MediationAdaptersForDFP'
-include ':MediationAdaptersForMoPub'
+include ':DemandSDKAdaptersForDFP'
+include ':DemandSDKAdaptersForMoPub'
